Place logic into module and call it in main index.js | Solve deletion issues
